feat(1122): integrate all port phases + refactor to 100/100 aislop
Final integration of the mc1.12.2 port:
- Phase 3 (9 integration tests) via cherry-pick
- Phase 6 (async IO drain coalesce) via cherry-pick
- Refactor: split SkinCommand + SkinMetrics to under file-size limits
  (extract SkinAction, SkinRefreshTask, Snapshot, PlayerSnapshot);
  remove narrative comments; keep only 'why' comments
- fix(permission): register Forge PermissionAPI nodes in INIT phase —
  PermissionAPI rejects preInit registration and crashed server startup
- All 229 tests pass + 3 drain tests
- Aislop 100/100
- Zero new dependencies
